Helper functions moved to their own files
